Version 4.2.3
Note: The #duplicateAllControlAndAltKeys preference, set true by default on Squeak 3.10, is a mis-feature that causes <ctl>C to be reported as not a control pressed event. Turn this preference off is you want command shells to be correctly interrupted with a <ctl>C character input from keyboard.

Change <ctl>C and <ctl>D event detection to not interpret <cmd> as the same thing as <ctl>. There was at least one combination of VM and and image for which the latter made sense, but it breaks cut and paste in the command shell window, which is not acceptable. For platforms/images in which <ctl>C gets reported as <cmd> event, this will just have to be treated as a bug.