CommandShell 4.5.6
When processing a command in a CommandShell, inform any dependents to display the command line. If the command came from a transcript stream that already contains the command line, do not update it. This allows a command shell that is controlled programmatically to display the command line input in a shell window if the shell window is open.