CommandShell 4.6.10
RemoteTask (poor man's multiprocessing): Add support for the Magma serializer. A RemoteTask can communicate its results to the parent image using any of the supported serializers (#referenceStream, #fuel, or #magma). See the class initialization methods for RemoteTask.

A RemoteTask represents a block to be evaluated in a headless Squeak child process. The results of evaluating the block are returned to the sender through a serialization stream on an OS pipe. Asynchronous event notification may be used to signal completion of remote processing.