CommandShell 4.6.13
When forking an image with PipeableOSProcess, close unused pipe handles for the connecting pipes immediately in both child and parent, otherwise testing #atEndOfFile can fail in one image beccause of the open handle in the other image.

In tests, add a #flush because C runtime may require it. Use upToEndOfFile and remove the now unneeded delay.