Update tests to document incorrect handling of standard input on oscog VMs Name: CommandShell-dtl.99 Time: 19 June 2019, 3:51:19.739277 pm Ancestors: CommandShell-dtl.98
CommandShell 4.7.3
Update tests to document incorrect handling of standard input on oscog
VMs, which breaks object serialization for pipeline connected VMs on Unix.
Document this with a test case, and update other test cases to fail
gracefully under this condition.

The failure occurs in <primitive: 'primitiveFileRead' module: 'FilePlugin'>
which populates only the first byte of the provided byte array when reading
from a standard input stream on oscog VMs. This may be a mis-feature
intended to accomodate Windows console input (speculation) but wrong
for the general case.

For reference, the interpreter VM does not exhibit this behavior, all tests pass.